pyproject.toml
[project]
name = "hackcheck"
version = "2.0.0"
description = "Official API wrapper for hackcheck.io"
authors = [{ name = "HackCheck", email = "support@hackcheck.io" }]
dependencies = ["httpx>=0.26.0", "pyserde>=0.12.7"]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
package-type = "library"
